Update getNowPlaying function
getNowPlaying function now returns a response from spotify, using the access token we pass to it from earlier in the script. Currently only outputs the entire body of the response, and doesn't do anything meaningful with the data (log it to a file etc). The current goal is to export the uri of the song provided in the response, and then dynamically create an embed on a page