Made a mess
Relevant information now logs to the node console, but I need to figure out a practical way to handle this data and place it on a webpage. I largely dislike the way i am using express.js and routing everything, too messy and I will need to refactor and move a lot of things.